Move getType from jest-matcher-utils to separate package
