+++ packages/jest-config/package.json
@@ -12,6 +12,7 @@
     "glob": "^7.1.1",
     "jest-environment-jsdom": "^20.0.3",
     "jest-environment-node": "^20.0.3",
+    "jest-get-type": "^20.0.3",
     "jest-jasmine2": "^20.0.4",
     "jest-matcher-utils": "^20.0.3",
     "jest-regex-util": "^20.0.3",
