+++ packages/jest-diff/package.json
@@ -12,6 +12,7 @@
     "chalk": "^1.1.3",
     "diff": "^3.2.0",
     "jest-matcher-utils": "^20.0.3",
+    "jest-get-type": "^20.0.3",
     "pretty-format": "^20.0.3"
   }
 }
