+++ packages/jest-matcher-utils/src/__tests__/index-test.js
@@ -10,13 +10,7 @@
 
 'use strict';
 
-const {
-  stringify,
-  getType,
-  ensureNumbers,
-  pluralize,
-  ensureNoExpected,
-} = require('../');
+const {stringify, ensureNumbers, pluralize, ensureNoExpected} = require('../');
 
 describe('.stringify()', () => {
   [
@@ -91,21 +85,6 @@ describe('.stringify()', () => {
     expect(stringify(big)).toMatchSnapshot();
     expect(stringify(small)).toMatchSnapshot();
   });
-});
-
-describe('.getType()', () => {
-  test('null', () => expect(getType(null)).toBe('null'));
-  test('undefined', () => expect(getType(undefined)).toBe('undefined'));
-  test('object', () => expect(getType({})).toBe('object'));
-  test('array', () => expect(getType([])).toBe('array'));
-  test('number', () => expect(getType(1)).toBe('number'));
-  test('string', () => expect(getType('oi')).toBe('string'));
-  test('function', () => expect(getType(() => {})).toBe('function'));
-  test('boolean', () => expect(getType(true)).toBe('boolean'));
-  test('symbol', () => expect(getType(Symbol.for('a'))).toBe('symbol'));
-  test('regexp', () => expect(getType(/abc/)).toBe('regexp'));
-  test('map', () => expect(getType(new Map())).toBe('map'));
-  test('set', () => expect(getType(new Set())).toBe('set'));
 });
 
 describe('.ensureNumbers()', () => {
