+++ packages/jest-matcher-utils/src/index.js
@@ -9,6 +9,7 @@
  */
 
 const chalk = require('chalk');
+const getType = require('jest-get-type');
 const prettyFormat = require('pretty-format');
 const {
   AsymmetricMatcher,
@@ -20,20 +21,6 @@ const {
 const PLUGINS = [AsymmetricMatcher, ReactElement, HTMLElement].concat(
   Immutable,
 );
-
-export type ValueType =
-  | 'array'
-  | 'boolean'
-  | 'function'
-  | 'null'
-  | 'number'
-  | 'object'
-  | 'regexp'
-  | 'map'
-  | 'set'
-  | 'string'
-  | 'symbol'
-  | 'undefined';
 
 const EXPECTED_COLOR = chalk.green;
 const EXPECTED_BG = chalk.bgGreen;
@@ -56,40 +43,6 @@ const NUMBERS = [
   'twelve',
   'thirteen',
 ];
-
-// get the type of a value with handling the edge cases like `typeof []`
-// and `typeof null`
-const getType = (value: any): ValueType => {
-  if (typeof value === 'undefined') {
-    return 'undefined';
-  } else if (value === null) {
-    return 'null';
-  } else if (Array.isArray(value)) {
-    return 'array';
-  } else if (typeof value === 'boolean') {
-    return 'boolean';
-  } else if (typeof value === 'function') {
-    return 'function';
-  } else if (typeof value === 'number') {
-    return 'number';
-  } else if (typeof value === 'string') {
-    return 'string';
-  } else if (typeof value === 'object') {
-    if (value.constructor === RegExp) {
-      return 'regexp';
-    } else if (value.constructor === Map) {
-      return 'map';
-    } else if (value.constructor === Set) {
-      return 'set';
-    }
-    return 'object';
-    // $FlowFixMe https://github.com/facebook/flow/issues/1015
-  } else if (typeof value === 'symbol') {
-    return 'symbol';
-  }
-
-  throw new Error(`value of unknown type: ${value}`);
-};
 
 const stringify = (object: any, maxDepth?: number = 10): string => {
   const MAX_LENGTH = 10000;
@@ -211,7 +164,6 @@ module.exports = {
   ensureExpectedIsNumber,
   ensureNoExpected,
   ensureNumbers,
-  getType,
   highlightTrailingWhitespace,
   matcherHint,
   pluralize,
