+++ packages/jest-validate/package.json
@@ -10,7 +10,7 @@
   "browser": "build-es5/index.js",
   "dependencies": {
     "chalk": "^1.1.3",
-    "jest-matcher-utils": "^20.0.3",
+    "jest-get-type": "^20.0.3",
     "leven": "^2.1.0",
     "pretty-format": "^20.0.3"
   }
